fix(profiles): add recovery buttons to lock dialogs and fix grey screen issue
